pyproject.toml
[tool.poetry]
name = "python-fastapi-template"
version = "0.1.0"
description = ""
authors = ["Takaaki Mizuno <takaaki.mizuno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
click = "^8.1.7"
psycopg2 = "^2.9.9"
bcrypt = "^4.1.3"
injector = "^0.21.0"
starlette = "^0.37.2"
pydantic-settings = "^2.3.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
yapf = "^0.40.2"
toml = "^0.10.2"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
